requirements.txt
fastapi
uvicorn
supabase
pydantic-settings
python-dotenv
openai
openpyxl